+++ ktlint-ruleset-standard/src/test/kotlin/com/pinterest/ktlint/ruleset/standard/FilenameRuleTest.kt
@@ -1,12 +1,12 @@
 package com.pinterest.ktlint.ruleset.standard
 
-import com.pinterest.ktlint.test.KtLintAssertThat.Companion.assertThat
+import com.pinterest.ktlint.test.KtLintAssertThat.Companion.assertThatRule
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
 
 class FilenameRuleTest {
-    private val fileNameRuleAssertThat = FilenameRule().assertThat()
+    private val fileNameRuleAssertThat = assertThatRule { FilenameRule() }
 
     @Test
     fun `Given a kotlin script file then ignore it`() {
@@ -56,7 +56,7 @@ class FilenameRuleTest {
             .hasLintViolationWithoutAutoCorrect(1, 1, "File name '$NON_PASCAL_CASE_NAME' should conform PascalCase")
     }
 
-    @ParameterizedTest(name = "Class type: {0}, expected result: {1}")
+    @ParameterizedTest(name = "Class type: {0}")
     @ValueSource(
         strings = [
             "class Foo",
@@ -81,12 +81,10 @@ class FilenameRuleTest {
     @ValueSource(
         strings = [
             "object Foo",
-            "typealias Foo = String",
-            "fun String.foo() = {}",
-            "fun foo() = {}"
+            "typealias Foo = String"
         ]
     )
-    fun `Given a file containing one top level declaration (no class or property) then the file should be named after the identifier`(
+    fun `Given a file containing one top level declaration then the file should be named after the identifier`(
         code: String
     ) {
         fileNameRuleAssertThat(code)
@@ -98,10 +96,13 @@ class FilenameRuleTest {
     @ValueSource(
         strings = [
             "val foo",
-            "const val FOO"
+            "const val FOO",
+            "fun String.foo() = {}",
+            "fun foo() = {}",
+            "operator fun Foo.plus(other: Foo): Foo { /* ... */ }"
         ]
     )
-    fun `Given a file containing one top level property declaration (non-private) then the file should conform to PascalCase`(
+    fun `Given a file containing one top level then the file should conform to PascalCase`(
         code: String
     ) {
         fileNameRuleAssertThat(code)
@@ -171,6 +172,14 @@ class FilenameRuleTest {
         fileNameRuleAssertThat(code)
             .asFileWithPath(UNEXPECTED_FILE_NAME)
             .hasLintViolationWithoutAutoCorrect(1, 1, "File '$UNEXPECTED_FILE_NAME' contains a single class and possibly also extension functions for that class and should be named same after that class 'Foo.kt'")
+    }
+
+    @Test
+    fun `Issue 1530 - Given a file which name should match PascalCase then this name may also contain letters with diacritics`() {
+        val code = "// some code"
+        fileNameRuleAssertThat(code)
+            .asFileWithPath("ŸëšThïsĮsÂllòwed123.kt")
+            .hasNoLintViolations()
     }
 
     private companion object {