+++ ktlint-ruleset-standard/src/test/kotlin/com/pinterest/ktlint/ruleset/standard/NoEmptyFirstLineInMethodBlockRuleTest.kt
@@ -1,11 +1,11 @@
 package com.pinterest.ktlint.ruleset.standard
 
-import com.pinterest.ktlint.test.KtLintAssertThat.Companion.assertThat
+import com.pinterest.ktlint.test.KtLintAssertThat.Companion.assertThatRule
 import com.pinterest.ktlint.test.LintViolation
 import org.junit.jupiter.api.Test
 
 class NoEmptyFirstLineInMethodBlockRuleTest {
-    private val noEmptyFirstLineInMethodBlockRuleAssertThat = NoEmptyFirstLineInMethodBlockRule().assertThat()
+    private val noEmptyFirstLineInMethodBlockRuleAssertThat = assertThatRule { NoEmptyFirstLineInMethodBlockRule() }
 
     @Test
     fun `Given a block in which the first line is not blank`() {
@@ -96,7 +96,7 @@ class NoEmptyFirstLineInMethodBlockRuleTest {
             fun fooBuilder() = object : Foo {
 
                 override fun foo() {
-                    TODO()
+                    bar()
                 }
             }
             """.trimIndent()